First working version of the app: let's call it an alpha?
We now have something vaguely resembling a blog. The index displays all the articles, clicking an article title takes you to a shoddily-formatted article page with a url corresponding to the article markdown filename.

UX still needs a lot of work, but that's all little fiddles. The app overall could probably benefit from some significant refactoring, but now we're at a point where it makes sense to start doing that.