Transparently handling undefined objects. Fixes #7
By 'transparently' I mean the template just skips them. This isn't 100% ideal, but for as long as the article loading function is returning undefined article objects, this is as close as I'm going to get to a definitive fix.

Interestingly enough, it looks like it's the *first* article that isn't getting loaded on time, not the last. I guess this teaches an important lesson about node's async bs: just because we're done with the 139th article, doesn't mean the first is finished.